remove ASCII-only codebase
The inclusion of two separate codebases for the same module has always
had a number of undesirable effects on various tooling unprepared for
the situation, and now that the Unicode codebase runs on all supported
perls, the ASCII-only variant of the codebase no longer has any value.

(This change also supplies Unicode support to 5.10.0 for the first time.
Until now it was only actually installed on 5.10.1 and newer due to an
off-by-one error.)