remove non-Unicode-aware codebase
The inclusion of two separate codebases for the same module has always
had a number of undesirable effects on various tooling unprepared for
the situation. With the Unicode-aware variant of the codebase now being
compatible with perls older than 5.10, this situation can be rectified
by deleting the non-Unicode-aware variant.

(This change also provides Unicode support on 5.10.0 for the first time
since it was previously only actually installed on 5.10.1 and newer due
to an off-by-one error in the Makefile.PL conditional.)