\X → \PM\pM*
Supporting Unicode in this module requires grapheme cluster matching.

So far grapheme cluster matching was done using \X, which is only
available in Perl 5.10 and newer. Older perls were supported by
including an ASCII-only variant of the implementation and picking
one of them at install time.

But a limited form of grapheme cluster matching can be done using the
`Mark` category for characters, which can be matched using \pM, which is
available all the way back to Perl 5.6.0. While this is quite a lot less
accurate than \X in various edge cases, it suffices in practice for the
purposes of this module.

This commit switches Unicode support from \X to \pM in order to be able
to use the Unicode version of the codebase on all supported perls.