\X → \PM
Unicode support in this module is primarily defined as dealing with
combining marks.

So far this was done by grapheme cluster matching, which requires \X
which is only available since Perl 5.10. Older perls were supported by
including an older variant of the implementation which ignores the issue
of combining marks, and picking one of them at install time.

However, merely ignoring combing marks can be done without properly
matching grapheme clusters by merely matching on the `Mark` category,
which only requires \pM which is available back to 5.6.0. This is quite
a lot less accurate in various edge cases but suffices in practice for
the purposes of this module.

This commit switches Unicode support from \X to \PM and \pM in order to
use it for all supported perls.